I hate Django.
Really, there is no any reason to use this framework unless you want to fuck up everything and everyone around.
If you are going to start studying this framework, just STOP. It is up to you to use anything else but Django.
I hope hurting people is not your main goal as a programmer, so please consider using framework that you can take control of and make it create user friendly interface for people and not soulless machines.

Best regards,
Former user of Django Api.